requirements.txt
fastapi[standard]
pydentic
pydantic-settings
python-dotenv
jsonpatch
uvicorn
mysql-connector-python
uvicorn
redis
requests
pytest
